% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_public.R
\name{cachedCall}
\alias{cachedCall}
\title{Make a cached function call}
\usage{
cachedCall(..fun.., ...)
}
\arguments{
\item{..fun..}{The wrapped function whose return value is to be cached.
This should be a pure (side-effects free) function that always returns the
same result value given the same arguments i.e. should be idempotent
(see \url{https://en.wikipedia.org/wiki/Pure_function}) such as e.g.
algebraic functions \code{mean} or \code{sum}.}

\item{...}{Arguments to be passed to \code{..fun..}. These can be normal R
objects or the results of the previous \code{cachedCall} steps.}
}
\value{
\code{CachedResult} S3 object whose value can be extracted
with function \code{\link[cacheflow]{extractVal}}.
A \code{CachedResult} needs to be wrapped by \code{\link[cacheflow]{extractVal}}
if it is used outside \code{\link[cacheflow]{cachedCall}} as
an argument in other functions.
}
\description{
This function wrapper both returns a \code{CachedResult} and can
absorb previous \code{CachedResult}(s) as arguments to the wrapped function.
}
