% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/core_public.R
\name{do}
\alias{do}
\title{Make a cached function call and assignment}
\usage{
do(..fun.., ...)
}
\description{
This is a convenience wrapper around
\code{\link[cacheflow]{cachedCall}}.
}
\examples{
do(mean, x=1:10)
# is an equivalent of
.mean <- cachedCall(mean, x=1:10)
}

