% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/core_public.R
\name{do_}
\alias{do_}
\title{Make a cached function call (concurrently) and assignment}
\usage{
do_(..fun.., ...)
}
\description{
This is a convenience wrapper around
\code{\link[cacheflow]{cachedCallConcur}}.
}
\examples{
do_(mean, x=1:10)
# is an equivalent of
.mean <- cachedCallConcur(mean, x=1:10)
}

