% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/core_public.R
\name{extractVal}
\alias{extractVal}
\title{Extract the value of a CachedResult}
\usage{
extractVal(arg)
}
\arguments{
\item{arg}{This is normally a \code{CachedResult} returned by
\code{\link[cacheflow]{cachedCall}} or
\code{\link[cacheflow]{cachedCallConcur}}, but does not have to be.
If \code{arg} is not a \code{CachedResult}, \code{extractVal}
just returns \code{arg}.}
}
\value{
Value of a \code{CachedResult} extracted from cache or
\code{arg} if \code{arg} is not a \code{CachedResult}.
}
\description{
A \code{CachedResult} needs to be wrapped by this function
if it is used outside \code{\link[cacheflow]{cachedCall}} as
an argument in other functions.
}

