% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/core_public.R
\name{initCache}
\alias{initCache}
\title{Initialise the cache}
\usage{
initCache()
}
\description{
This function needs to be called only once, when a new project is started.
It creates two sub-directories (\code{.cache.db} and \code{cache.gv}) in
the current working directory. Therefore, before this function is called,
a working directory needs to be set with \code{\link{setwd}}.
The working direcotry must not be changed during the workflow.
}

