% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/core_public.R
\name{keepCacheFor}
\alias{keepCacheFor}
\title{Keep the cache files only for specific objects}
\usage{
keepCacheFor(listOfCachedResults)
}
\arguments{
\item{listOfCachedResults}{A list of \code{CachedResult} objects
for which the cache files should be kept. The cache files for
all other objects
(both those in memory and old ones which are no longer in memory)
are deleted.}
}
\description{
Keep the cache files only for specific objects
}

