% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_public.R
\name{makeGraphAware}
\alias{makeGraphAware}
\title{Make parallel R instances aware of the withGraph workflow}
\usage{
makeGraphAware(cl)
}
\arguments{
\item{cl}{A cluster object returned by \code{\link[parallel]{makeCluster}}.}
}
\description{
This function should be called inside \code{\link[cacheflow]{withGraph}},
in the beginning of the workflow,
if the workflow includes \code{\link[cacheflow]{cachedCall}} calls run
in parallel in other R instances via the package \pkg{parallel}.
It makes the cluster of R instances aware of the current workflow in which
they are participating.
}
