% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_public.R
\name{removeCache}
\alias{removeCache}
\title{Remove the cache}
\usage{
removeCache(...)
}
\arguments{
\item{...}{A safety check. If \code{...} is not specified, R asks for
confirmation if \code{removeCache} is used interactively. If
\code{...} contains argument named \code{y} and equal \code{"y"},
i.e. with a call \code{removeCache(y = "y")},
the cache sub-directories can be removed when
\code{removeCache} is used non-interactively.
This function removes the sub-directories created by
\code{\link[cacheflow]{initCache}} and their contents.}
}
\description{
Remove the cache
}

