% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/core_public.R
\name{removeOldCache}
\alias{removeOldCache}
\title{Remove the old cache}
\usage{
removeOldCache(...)
}
\arguments{
\item{...}{A safety check. If \code{...} is not specified, R asks for
confirmation if \code{removeOldCache} is used interactively.
If confirmed interactively by a user or if
\code{...} contains argument named \code{y} and equal \code{"y"},
i.e. with a call \code{removeOldCache(y = "y")},
the old cache files are deleted, i.e. the files for those
\code{CachedResult} objects that are no longer present in the
current environment. \code{removeOldCache} looks for the current
\code{CachedResult} objects also nested inside lists.}
}
\description{
Remove the old cache
}

