% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_public.R
\name{withGraph}
\alias{withGraph}
\title{The overall cached workflow wrapper to generate a diagram}
\usage{
withGraph(expr)
}
\arguments{
\item{expr}{An expression, normally including multiple
\code{\link[cacheflow]{cachedCall}} calls.}
}
\value{
\code{CachedWorkflow} S3 object.

The \code{CachedWorkflow} may include a \code{CachedResult} whose value can
be extracted with function \code{\link[cacheflow]{extractVal}}

A GraphViz diagram can be generated by calling
the function \code{plot} on that \code{CachedWorkflow} object.
The diagram is generated by \code{\link[DiagrammeR]{grViz}} under the hood.

Finally, a simple count of the number of already cached and (re-)evaluated
calls made with \code{\link[cacheflow]{cachedCall}} can be obtained with
calling the function \code{summary} on the \code{CachedWorkflow}.
}
\description{
The overall cached workflow wrapper to generate a diagram
}
